LoginDocente, NavbarDocente, HeederDocente, y re direccion de crear curso y tus cursos
