+++ estu_prime/src/pages/Home/crearCuentaEstudiante.jsx
@@ -1,10 +1,10 @@
-
+import { Fragment } from 'react';
 import FormE from '../../components/form/formE';
 const crearCuentaEstudiante = () => {
   return (
-    <div>
+    <Fragment>
       <FormE/>
-    </div> 
+    </Fragment> 
   )
 }
