+++ estu_prime/src/pages/LoginDoc/inicioDoc.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function inicioDoc() {
+  return (
+    <div>inicioDoc</div>
+  )
+}
+
+export default inicioDoc
